refactor(factory): #326 split factory_test.go along review-named seams (pure move)
