+++ factory/factory_test.go
@@ -9,16 +9,10 @@ import (
 	"testing"
 	"time"
 
-	"github.com/lychee-technology/forma/internal/federated"
-	"github.com/lychee-technology/forma/internal/schemameta"
-
 	"github.com/jackc/pgx/v5/pgxpool"
 	"github.com/lychee-technology/forma"
-	"github.com/pashagolub/pgxmock/v4"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"go.uber.org/zap"
-	"go.uber.org/zap/zaptest/observer"
 )
 
 // ---------------------------------------------------------------------------
@@ -141,384 +135,6 @@ func createTempTables(t *testing.T, ctx context.Context, pool *pgxpool.Pool) (sc
 	})
 
 	return schemaRegistryTable, entityMainTable, eavDataTable
-}
-
-// ---------------------------------------------------------------------------
-// Unit tests for collectTablesFromPool (uses pgxmock)
-// ---------------------------------------------------------------------------
-
-func TestCollectTablesFromPool_QueryError(t *testing.T) {
-	mock, err := pgxmock.NewPool()
-	require.NoError(t, err)
-	defer mock.Close()
-
-	mock.ExpectQuery(`SELECT table_name FROM information_schema.tables`).
-		WithArgs("tenant_schema").
-		WillReturnError(assert.AnError)
-
-	_, err = collectTablesFromPool(context.Background(), mock, "tenant_schema")
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "failed to verify database connection")
-	require.NoError(t, mock.ExpectationsWereMet())
-}
-
-func TestCollectTablesFromPool_Success(t *testing.T) {
-	mock, err := pgxmock.NewPool()
-	require.NoError(t, err)
-	defer mock.Close()
-
-	rows := pgxmock.NewRows([]string{"table_name"}).
-		AddRow("schema_registry").
-		AddRow("eav_data")
-	mock.ExpectQuery(`SELECT table_name FROM information_schema.tables`).
-		WithArgs("custom_schema").
-		WillReturnRows(rows)
-
-	tables, err := collectTablesFromPool(context.Background(), mock, "custom_schema")
-	require.NoError(t, err)
-	assert.Contains(t, tables, "schema_registry")
-	assert.Contains(t, tables, "eav_data")
-	require.NoError(t, mock.ExpectationsWereMet())
-}
-
-func TestCollectTablesFromPool_DefaultSchema(t *testing.T) {
-	mock, err := pgxmock.NewPool()
-	require.NoError(t, err)
-	defer mock.Close()
-
-	rows := pgxmock.NewRows([]string{"table_name"}).
-		AddRow("schema_registry").
-		AddRow("eav_data")
-	mock.ExpectQuery(`SELECT table_name FROM information_schema.tables`).
-		WithArgs("public").
-		WillReturnRows(rows)
-
-	_, err = collectTablesFromPool(context.Background(), mock, "")
-	require.NoError(t, err)
-	require.NoError(t, mock.ExpectationsWereMet())
-}
-
-// ---------------------------------------------------------------------------
-// Unit tests for NewEntityManagerWithConfig (uses test hooks + mock loader)
-// ---------------------------------------------------------------------------
-
-type mockMetadataLoader struct {
-	cache *schemameta.MetadataCache
-	err   error
-}
-
-func (m *mockMetadataLoader) LoadMetadata(ctx context.Context) (*schemameta.MetadataCache, error) {
-	return m.cache, m.err
-}
-
-func unitEntityManagerDeps(cache *schemameta.MetadataCache) entityManagerDependencies {
-	deps := defaultEntityManagerDependencies()
-	deps.collectTables = func(ctx context.Context, pool queryPool, schema string) ([]string, error) {
-		return []string{"schema_registry", "eav_data", "entity_main"}, nil
-	}
-	deps.newMetadataLoader = func(pool *pgxpool.Pool, schemaTable, schemaDir string) metadataLoader {
-		return &mockMetadataLoader{cache: cache, err: nil}
-	}
-	return deps
-}
-
-// unitEntityManagerConfig is the config half of the successful-construction
-// fixture: the table names unitEntityManagerDeps' collector reports, plus an
-// empty schema directory.
-func unitEntityManagerConfig(t *testing.T) *forma.Config {
-	t.Helper()
-	config := forma.DefaultConfig(newMockSchemaRegistry())
-	config.Database.TableNames = forma.TableNames{
-		SchemaRegistry: "schema_registry",
-		EAVData:        "eav_data",
-		EntityMain:     "entity_main",
-	}
-	config.Entity.SchemaDirectory = t.TempDir()
-	return config
-}
-
-func TestNewEntityManagerWithConfig_Unit_TableCollectorError(t *testing.T) {
-	t.Parallel()
-
-	deps := defaultEntityManagerDependencies()
-	deps.collectTables = func(ctx context.Context, pool queryPool, schema string) ([]string, error) {
-		return nil, assert.AnError
-	}
-
-	config := forma.DefaultConfig(newMockSchemaRegistry())
-
-	em, err := newEntityManagerWithConfigContext(context.Background(), config, nil, deps)
-
-	assert.Nil(t, em)
-	assert.Error(t, err)
-}
-
-func TestNewEntityManagerWithConfig_Unit_MissingRequiredTables(t *testing.T) {
-	t.Parallel()
-
-	deps := defaultEntityManagerDependencies()
-	deps.collectTables = func(ctx context.Context, pool queryPool, schema string) ([]string, error) {
-		return []string{"schema_registry"}, nil
-	}
-
-	config := forma.DefaultConfig(newMockSchemaRegistry())
-	config.Database.TableNames = forma.TableNames{
-		SchemaRegistry: "schema_registry",
-		EAVData:        "eav_data",
-		EntityMain:     "entity_main",
-	}
-
-	em, err := newEntityManagerWithConfigContext(context.Background(), config, nil, deps)
-
-	assert.Nil(t, em)
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "required tables are missing")
-}
-
-func TestNewEntityManagerWithConfig_Unit_MissingEntityMainTable(t *testing.T) {
-	t.Parallel()
-
-	deps := defaultEntityManagerDependencies()
-	deps.collectTables = func(ctx context.Context, pool queryPool, schema string) ([]string, error) {
-		// schema_registry and eav_data present, but entity_main is absent
-		return []string{"schema_registry", "eav_data"}, nil
-	}
-
-	config := forma.DefaultConfig(newMockSchemaRegistry())
-	config.Database.TableNames = forma.TableNames{
-		SchemaRegistry: "schema_registry",
-		EAVData:        "eav_data",
-		EntityMain:     "entity_main",
-	}
-
-	em, err := newEntityManagerWithConfigContext(context.Background(), config, nil, deps)
-
-	assert.Nil(t, em)
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "required tables are missing")
-}
-
-func TestNewEntityManagerWithConfig_Unit_MetadataLoaderError(t *testing.T) {
-	t.Parallel()
-
-	cache := schemameta.NewMetadataCache()
-	deps := unitEntityManagerDeps(cache)
-	deps.newMetadataLoader = func(pool *pgxpool.Pool, schemaTable, schemaDir string) metadataLoader {
-		return &mockMetadataLoader{cache: cache, err: fmt.Errorf("simulated loader error")}
-	}
-
-	config := forma.DefaultConfig(newMockSchemaRegistry())
-	config.Database.TableNames = forma.TableNames{
-		SchemaRegistry: "schema_registry",
-		EAVData:        "eav_data",
-		EntityMain:     "entity_main",
-	}
-	config.Entity.SchemaDirectory = t.TempDir()
-
-	em, err := newEntityManagerWithConfigContext(context.Background(), config, nil, deps)
-
-	assert.Nil(t, em)
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "failed to load metadata")
-}
-
-func TestNewEntityManagerWithConfig_Unit_NilSchemaRegistry(t *testing.T) {
-	t.Parallel()
-
-	cache := schemameta.NewMetadataCache()
-	deps := unitEntityManagerDeps(cache)
-
-	config := forma.DefaultConfig(nil)
-	config.Database.TableNames = forma.TableNames{
-		SchemaRegistry: "schema_registry",
-		EAVData:        "eav_data",
-		EntityMain:     "entity_main",
-	}
-	config.Entity.SchemaDirectory = t.TempDir()
-
-	em, err := newEntityManagerWithConfigContext(context.Background(), config, nil, deps)
-
-	assert.Nil(t, em)
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "config.SchemaRegistry is required")
-}
-
-func TestNewEntityManagerWithConfig_Unit_Success(t *testing.T) {
-	t.Parallel()
-
-	cache := schemameta.NewMetadataCache()
-	deps := unitEntityManagerDeps(cache)
-
-	config := unitEntityManagerConfig(t)
-
-	em, err := newEntityManagerWithConfigContext(context.Background(), config, nil, deps)
-
-	assert.NotNil(t, em)
-	assert.NoError(t, err)
-}
-
-func TestNewEntityManagerWithConfig_Unit_SchemaQualifiedTableNames(t *testing.T) {
-	t.Parallel()
-
-	cache := schemameta.NewMetadataCache()
-	deps := unitEntityManagerDeps(cache)
-	deps.collectTables = func(ctx context.Context, pool queryPool, schema string) ([]string, error) {
-		assert.Equal(t, "tenant", schema)
-		return []string{"schema_registry", "eav_data", "entity_main"}, nil
-	}
-	deps.newMetadataLoader = func(pool *pgxpool.Pool, schemaTable, schemaDir string) metadataLoader {
-		assert.Equal(t, "tenant.schema_registry", schemaTable)
-		return &mockMetadataLoader{cache: cache, err: nil}
-	}
-
-	config := forma.DefaultConfig(newMockSchemaRegistry())
-	config.Database.Schema = "tenant"
-	config.Database.TableNames = forma.TableNames{
-		SchemaRegistry: "tenant.schema_registry",
-		EAVData:        "tenant.eav_data",
-		EntityMain:     "tenant.entity_main",
-	}
-	config.Entity.SchemaDirectory = t.TempDir()
-
-	em, err := newEntityManagerWithConfigContext(context.Background(), config, nil, deps)
-
-	assert.NotNil(t, em)
-	assert.NoError(t, err)
-}
-
-func TestNewEntityManagerWithConfig_Unit_SchemaParamQualifiesUnqualifiedTableNames(t *testing.T) {
-	t.Parallel()
-
-	cache := schemameta.NewMetadataCache()
-	deps := unitEntityManagerDeps(cache)
-	deps.collectTables = func(ctx context.Context, pool queryPool, schema string) ([]string, error) {
-		assert.Equal(t, "tenant", schema)
-		return []string{"schema_registry", "eav_data", "entity_main"}, nil
-	}
-	deps.newMetadataLoader = func(pool *pgxpool.Pool, schemaTable, schemaDir string) metadataLoader {
-		assert.Equal(t, "tenant.schema_registry", schemaTable)
-		return &mockMetadataLoader{cache: cache, err: nil}
-	}
-
-	config := forma.DefaultConfig(newMockSchemaRegistry())
-	config.Database.Schema = "tenant"
-	config.Database.TableNames = forma.TableNames{
-		SchemaRegistry: "schema_registry",
-		EAVData:        "eav_data",
-		EntityMain:     "entity_main",
-		ChangeLog:      "change_log",
-	}
-	config.Entity.SchemaDirectory = t.TempDir()
-
-	em, err := newEntityManagerWithConfigContext(context.Background(), config, nil, deps)
-
-	assert.NotNil(t, em)
-	assert.NoError(t, err)
-}
-
-func TestNewEntityManagerWithConfigContext_Unit_PropagatesContextToTableCollector(t *testing.T) {
-	t.Parallel()
-
-	cache := schemameta.NewMetadataCache()
-	ctx, cancel := context.WithCancel(context.Background())
-	cancel()
-
-	deps := unitEntityManagerDeps(cache)
-	deps.collectTables = func(ctx context.Context, pool queryPool, schema string) ([]string, error) {
-		assert.ErrorIs(t, ctx.Err(), context.Canceled)
-		return []string{"schema_registry", "eav_data", "entity_main"}, nil
-	}
-
-	config := forma.DefaultConfig(newMockSchemaRegistry())
-	config.Database.TableNames = forma.TableNames{
-		SchemaRegistry: "schema_registry",
-		EAVData:        "eav_data",
-		EntityMain:     "entity_main",
-	}
-	config.Entity.SchemaDirectory = t.TempDir()
-
-	em, err := newEntityManagerWithConfigContext(ctx, config, nil, deps)
-
-	assert.NotNil(t, em)
-	assert.NoError(t, err)
-}
-
-func TestNewEntityManagerWithConfigContext_Unit_PropagatesContextToDuckDBFactory(t *testing.T) {
-	t.Parallel()
-
-	cache := schemameta.NewMetadataCache()
-	ctx, cancel := context.WithCancel(context.Background())
-	cancel()
-
-	deps := unitEntityManagerDeps(cache)
-	deps.newDuckDBClient = func(ctx context.Context, cfg forma.DuckDBConfig) (*federated.DuckDBClient, error) {
-		assert.ErrorIs(t, ctx.Err(), context.Canceled)
-		return nil, context.Canceled
-	}
-
-	config := forma.DefaultConfig(newMockSchemaRegistry())
-	config.Database.TableNames = forma.TableNames{
-		SchemaRegistry: "schema_registry",
-		EAVData:        "eav_data",
-		EntityMain:     "entity_main",
-	}
-	config.Entity.SchemaDirectory = t.TempDir()
-	config.DuckDB.Enabled = true
-
-	em, err := newEntityManagerWithConfigContext(ctx, config, nil, deps)
-
-	assert.NotNil(t, em)
-	assert.NoError(t, err)
-}
-
-func TestNewDuckDBCircuitBreakerUsesDefaultParametersForZeroConfig(t *testing.T) {
-	t.Parallel()
-
-	breaker := newDuckDBCircuitBreaker(forma.DuckDBConfig{})
-
-	for i := 0; i < 4; i++ {
-		breaker.RecordFailure()
-		assert.False(t, breaker.IsOpen())
-	}
-	breaker.RecordFailure()
-	assert.True(t, breaker.IsOpen())
-}
-
-func TestNewDuckDBCircuitBreakerUsesConfiguredParameters(t *testing.T) {
-	t.Parallel()
-
-	breaker := newDuckDBCircuitBreaker(forma.DuckDBConfig{
-		CircuitBreakerFailureThreshold: 2,
-		CircuitBreakerWindow:           time.Minute,
-		CircuitBreakerOpenDuration:     time.Minute,
-	})
-
-	breaker.RecordFailure()
-	assert.False(t, breaker.IsOpen())
-	breaker.RecordFailure()
-	assert.True(t, breaker.IsOpen())
-}
-
-func TestNewDuckDBCircuitBreakerDoesNotWarnForDefaultThreshold(t *testing.T) {
-	core, logs := observer.New(zap.WarnLevel)
-	restore := zap.ReplaceGlobals(zap.New(core))
-	t.Cleanup(restore)
-
-	breaker := newDuckDBCircuitBreaker(forma.DefaultConfig(nil).DuckDB)
-
-	require.NotNil(t, breaker)
-	assert.Equal(t, 0, logs.FilterMessage("circuitBreakerThreshold is deprecated and ignored; use circuitBreakerFailureThreshold instead").Len())
-}
-
-func TestNewDuckDBCircuitBreakerWarnsForDeprecatedThreshold(t *testing.T) {
-	core, logs := observer.New(zap.WarnLevel)
-	restore := zap.ReplaceGlobals(zap.New(core))
-	t.Cleanup(restore)
-
-	breaker := newDuckDBCircuitBreaker(forma.DuckDBConfig{CircuitBreakerThreshold: 0.5})
-
-	require.NotNil(t, breaker)
-	assert.Equal(t, 1, logs.FilterMessage("circuitBreakerThreshold is deprecated and ignored; use circuitBreakerFailureThreshold instead").Len())
 }
 
 // ---------------------------------------------------------------------------
@@ -665,232 +281,5 @@ func TestNewEntityManagerWithConfig_NilPool(t *testing.T) {
 	// This will panic with nil pool - we verify the function doesn't handle nil gracefully
 	assert.Panics(t, func() {
 		_, _ = NewEntityManagerWithConfig(config, nil)
-	})
-}
-
-// ---------------------------------------------------------------------------
-// Integration Tests for NewFileSchemaRegistry
-// ---------------------------------------------------------------------------
-
-func TestNewFileSchemaRegistry_Integration_Success(t *testing.T) {
-	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-	defer cancel()
-
-	pool := connectTestPostgres(t, ctx)
-
-	// Create schema registry table
-	suffix := time.Now().UnixNano()
-	tableName := fmt.Sprintf("schema_registry_fsr_%d", suffix)
-
-	_, err := pool.Exec(ctx, fmt.Sprintf(`
-		CREATE TABLE %s (
-			schema_name TEXT PRIMARY KEY,
-			schema_id SMALLINT NOT NULL
-		)
-	`, tableName))
-	require.NoError(t, err)
-
-	t.Cleanup(func() {
-		cleanupCtx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-		_, _ = pool.Exec(cleanupCtx, fmt.Sprintf("DROP TABLE IF EXISTS %s", tableName))
-	})
-
-	// Insert schemas
-	_, err = pool.Exec(ctx, fmt.Sprintf(
-		"INSERT INTO %s (schema_name, schema_id) VALUES ($1, $2)",
-		tableName,
-	), "test", int16(1))
-	require.NoError(t, err)
-
-	// Create temp schema directory with required files
-	dir := t.TempDir()
-	writeAttributesFile(t, dir, "test", map[string]any{
-		"id": map[string]any{
-			"attributeID": float64(1),
-			"valueType":   "text",
-		},
-	})
-	writeSchemaFile(t, dir, "test", map[string]any{
-		"type": "object",
-		"properties": map[string]any{
-			"id": map[string]any{"type": "string"},
-		},
-	})
-
-	registry, err := NewFileSchemaRegistry(pool, tableName, dir)
-
-	assert.NoError(t, err)
-	assert.NotNil(t, registry)
-
-	// Verify registry works
-	schemas := registry.ListSchemas()
-	assert.Contains(t, schemas, "test")
-
-	schemaID, cache, err := registry.GetSchemaAttributeCacheByName("test")
-	assert.NoError(t, err)
-	assert.Equal(t, int16(1), schemaID)
-	assert.Contains(t, cache, "id")
-}
-
-func TestNewFileSchemaRegistry_Integration_EmptyRegistry(t *testing.T) {
-	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-	defer cancel()
-
-	pool := connectTestPostgres(t, ctx)
-
-	// Create empty schema registry table
-	suffix := time.Now().UnixNano()
-	tableName := fmt.Sprintf("schema_registry_empty_%d", suffix)
-
-	_, err := pool.Exec(ctx, fmt.Sprintf(`
-		CREATE TABLE %s (
-			schema_name TEXT PRIMARY KEY,
-			schema_id SMALLINT NOT NULL
-		)
-	`, tableName))
-	require.NoError(t, err)
-
-	t.Cleanup(func() {
-		cleanupCtx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-		_, _ = pool.Exec(cleanupCtx, fmt.Sprintf("DROP TABLE IF EXISTS %s", tableName))
-	})
-
-	// Don't insert any schemas
-
-	registry, err := NewFileSchemaRegistry(pool, tableName, t.TempDir())
-
-	assert.Nil(t, registry)
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "no schemas found")
-}
-
-func TestNewFileSchemaRegistry_Integration_InvalidDirectory(t *testing.T) {
-	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-	defer cancel()
-
-	pool := connectTestPostgres(t, ctx)
-
-	// Create schema registry table with a schema entry
-	suffix := time.Now().UnixNano()
-	tableName := fmt.Sprintf("schema_registry_invdir_%d", suffix)
-
-	_, err := pool.Exec(ctx, fmt.Sprintf(`
-		CREATE TABLE %s (
-			schema_name TEXT PRIMARY KEY,
-			schema_id SMALLINT NOT NULL
-		)
-	`, tableName))
-	require.NoError(t, err)
-
-	t.Cleanup(func() {
-		cleanupCtx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-		_, _ = pool.Exec(cleanupCtx, fmt.Sprintf("DROP TABLE IF EXISTS %s", tableName))
-	})
-
-	// Insert a schema
-	_, err = pool.Exec(ctx, fmt.Sprintf(
-		"INSERT INTO %s (schema_name, schema_id) VALUES ($1, $2)",
-		tableName,
-	), "test", int16(1))
-	require.NoError(t, err)
-
-	// Use non-existent directory
-	registry, err := NewFileSchemaRegistry(pool, tableName, "/nonexistent/directory")
-
-	assert.Nil(t, registry)
-	assert.Error(t, err)
-}
-
-func TestNewFileSchemaRegistry_Integration_MissingAttributesFile(t *testing.T) {
-	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-	defer cancel()
-
-	pool := connectTestPostgres(t, ctx)
-
-	// Create schema registry table
-	suffix := time.Now().UnixNano()
-	tableName := fmt.Sprintf("schema_registry_noattr_%d", suffix)
-
-	_, err := pool.Exec(ctx, fmt.Sprintf(`
-		CREATE TABLE %s (
-			schema_name TEXT PRIMARY KEY,
-			schema_id SMALLINT NOT NULL
-		)
-	`, tableName))
-	require.NoError(t, err)
-
-	t.Cleanup(func() {
-		cleanupCtx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-		_, _ = pool.Exec(cleanupCtx, fmt.Sprintf("DROP TABLE IF EXISTS %s", tableName))
-	})
-
-	// Insert a schema
-	_, err = pool.Exec(ctx, fmt.Sprintf(
-		"INSERT INTO %s (schema_name, schema_id) VALUES ($1, $2)",
-		tableName,
-	), "test", int16(1))
-	require.NoError(t, err)
-
-	// Create empty directory (no attributes file)
-	dir := t.TempDir()
-
-	registry, err := NewFileSchemaRegistry(pool, tableName, dir)
-
-	assert.Nil(t, registry)
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "failed to read attributes file")
-}
-
-func TestNewFileSchemaRegistry_Integration_DuplicateSchemaIDFails(t *testing.T) {
-	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-	defer cancel()
-
-	pool := connectTestPostgres(t, ctx)
-
-	suffix := time.Now().UnixNano()
-	tableName := fmt.Sprintf("schema_registry_dup_id_%d", suffix)
-
-	_, err := pool.Exec(ctx, fmt.Sprintf(`
-		CREATE TABLE %s (
-			schema_name TEXT PRIMARY KEY,
-			schema_id SMALLINT NOT NULL
-		)
-	`, tableName))
-	require.NoError(t, err)
-
-	t.Cleanup(func() {
-		cleanupCtx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-		_, _ = pool.Exec(cleanupCtx, fmt.Sprintf("DROP TABLE IF EXISTS %s", tableName))
-	})
-
-	_, err = pool.Exec(ctx, fmt.Sprintf(
-		"INSERT INTO %s (schema_name, schema_id) VALUES ($1, $2), ($3, $4)",
-		tableName,
-	), "alpha", int16(1), "beta", int16(1))
-	require.NoError(t, err)
-
-	dir := t.TempDir()
-	writeAttributesFile(t, dir, "alpha", map[string]any{
-		"id": map[string]any{"attributeID": float64(1), "valueType": "text"},
-	})
-	writeAttributesFile(t, dir, "beta", map[string]any{
-		"id": map[string]any{"attributeID": float64(2), "valueType": "text"},
-	})
-
-	registry, err := NewFileSchemaRegistry(pool, tableName, dir)
-	assert.Nil(t, registry)
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "duplicate schema id")
-}
-
-func TestNewFileSchemaRegistry_NilPool(t *testing.T) {
-	// This will panic with nil pool
-	assert.Panics(t, func() {
-		_, _ = NewFileSchemaRegistry(nil, "test_table", t.TempDir())
 	})
 }
